+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for creating applications and gateway with an organization as the initial owner in the Console.
 - Hide views and features in the Console that the user and stack configuration does not meet the necessary requirements for.
+- Full range of Join EUI prefixes in the Console.
 
 ### Changed